tests/benchmarks/bench_service/maas/requirements.txt
aiohttp~=3.7.4
Pillow==8.3.1
timm==0.4.12
torch==1.9.0
torchvision==0.10.0
pandas==1.3.1
docker~=5.0.0
aiodogstatsd~=0.15.0